Adding the UserRequestController
