+++ src/FairPlayTubeSln/FairPlayTubeSln.sln
@@ -60,6 +60,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FairPlayTube.Controllers", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FairPlayTube.Translator.Windows", "FairPlayTube.Translator.Windows\FairPlayTube.Translator.Windows.csproj", "{2FDD6BF6-C668-4DD2-AE61-0083511694CE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FairPlayTube.SharedConfiguration", "FairPlayTube.SharedConfiguration\FairPlayTube.SharedConfiguration.csproj", "{C9F45BF8-B4F2-4C1B-8ACF-4446128A100F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -128,6 +130,10 @@ Global
 		{2FDD6BF6-C668-4DD2-AE61-0083511694CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2FDD6BF6-C668-4DD2-AE61-0083511694CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2FDD6BF6-C668-4DD2-AE61-0083511694CE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C9F45BF8-B4F2-4C1B-8ACF-4446128A100F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C9F45BF8-B4F2-4C1B-8ACF-4446128A100F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C9F45BF8-B4F2-4C1B-8ACF-4446128A100F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C9F45BF8-B4F2-4C1B-8ACF-4446128A100F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
